ai_assistant/requirements.txt
# Personal AI Assistant Dependencies

# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# LangChain Ecosystem
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.0

# AI Models
ollama==0.1.7

# Vector Memory
chromadb==0.4.18

# Web Search
tavily-python==0.3.0

# Google APIs
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0

# Image Processing
torch==2.1.0
torchvision==0.16.0
clip-by-openai==1.0
Pillow==10.1.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==0.24.0